SFDM-3384 [Back][CX] Доработать метод PATCH /api​/cx​/v1​/bpmn/cj/{cj-id} заполнения атрибутов CJ корректной сортировкой

+++ src/main/java/ru/beeline/cxbackend/dto/BiByCjStepDto.java
@@ -7,9 +7,11 @@ package ru.beeline.cxbackend.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BiByCjStepDto {
     @JsonProperty("id_bi")
     private Long idBi;
-    private Long order;
+    private BigDecimal order;
 }
