+++ src/main/java/ru/beeline/cxbackend/dto/StepDtoV3.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,7 +21,7 @@ public class StepDtoV3 {
 
     private Long cjId;
 
-    private Integer order;
+    private BigDecimal order;
 
     private String name;
